backend/requirements.txt
fastapi==0.124.4
gensim==4.4.0
jieba==0.42.1
numpy==1.26.4
pdfplumber==0.11.8
peft==0.10.0
pydantic==2.12.5
python_docx==1.2.0
python==3.12.4
scikit_learn==1.4.2
scipy==1.13.1
torch==2.5.0
transformers==4.41.2
sentence_transformers==2.6.1

